requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
redis
bcrypt
PyJWT
pydantic
pydantic-settings
pytest
